Show disclaimer first on site entry with yes/no choices
